app/requirements.txt
alembic==1.4.3
amqp==5.0.2
attrs==20.2.0
bcrypt==3.2.0
billiard==3.6.3.0
celery==5.0.4
cffi==1.14.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
cryptography==3.2.1
ecdsa==0.14.1
fastapi==0.61.0
h11==0.9.0
httptools==0.1.1
injector==0.18.4
jinja2==2.11.2
kombu==5.0.2
Mako==1.1.3
MarkupSafe==1.1.1
passlib==1.7.4
prompt-toolkit==3.0.8
punq==0.4.1
psycopg2==2.8.6
pyasn1==0.4.8
pycparser==2.20
pydantic==1.6.1
python-dateutil==2.8.1
python-editor==1.0.4
python-jose==3.2.0
python-multipart==0.0.5
pytz==2020.4
redis==3.5.3
rsa==4.6
six==1.15.0
SQLAlchemy==1.3.19
starlette==0.13.6
typing-extensions==3.7.4.3
uvicorn==0.11.8
uvloop==0.14.0
vine==5.0.0
wcwidth==0.2.5
websockets==8.1
